выносит фнкции валидации форм в отдельный файл
